Create .kt
Explanation:

Two variables x and y are defined with values 10 and 20, respectively.
A simple IF statement checks whether x is less than y, and if so, prints a message to the console.
An IF-ELSE statement checks whether x is greater than y, and if so, prints one message to the console. Otherwise, it prints a different message to the console.
An IF-ELSE-IF statement checks whether x is greater than y, equal to y, or less than y, and prints a different message to the console for each case.